EBEGU-1916: EWK-Abfrage für Jurist ermöglichen, GS soll Mutation machen können auch wenn Gesuch bei STV, Gesuche und Mitteilungen löschen muss auch für Schulamt möglich sein

+++ ebegu-server/src/main/java/ch/dvbern/ebegu/services/EbeguVorlageServiceBean.java
@@ -61,8 +61,6 @@ import ch.dvbern.lib.cdipersistence.Persistence;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.Validate;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import static ch.dvbern.ebegu.enums.UserRoleName.ADMIN;
 import static ch.dvbern.ebegu.enums.UserRoleName.SUPER_ADMIN;
@@ -73,8 +71,6 @@ import static ch.dvbern.ebegu.enums.UserRoleName.SUPER_ADMIN;
 @Stateless
 @Local(EbeguVorlageService.class)
 public class EbeguVorlageServiceBean extends AbstractBaseService implements EbeguVorlageService {
-
-	private static final Logger LOGGER = LoggerFactory.getLogger(EbeguVorlageServiceBean.class.getSimpleName());
 
 	@Inject
 	private Persistence persistence;
@@ -198,8 +194,9 @@ public class EbeguVorlageServiceBean extends AbstractBaseService implements Ebeg
 		EbeguVorlage ebeguVorlageEntity = ebeguVorlage.orElseThrow(() -> new EbeguEntityNotFoundException
 			("removeEbeguVorlage", ErrorCodeEnum.ERROR_ENTITY_NOT_FOUND, id));
 
-		fileSaverService.remove(ebeguVorlageEntity.getVorlage().getFilepfad());
-
+		if (ebeguVorlageEntity.getVorlage() != null) {
+			fileSaverService.remove(ebeguVorlageEntity.getVorlage().getFilepfad());
+		}
 		ebeguVorlageEntity.setVorlage(null);
 		updateEbeguVorlage(ebeguVorlageEntity);
 	}
@@ -255,8 +252,12 @@ public class EbeguVorlageServiceBean extends AbstractBaseService implements Ebeg
 
 	@Override
 	@PermitAll
+	@Nullable
 	public Vorlage getBenutzerhandbuch() {
 		UserRole userRole = principalBean.discoverMostPrivilegedRole();
+		if (userRole == null) {
+			return null;
+		}
 		EbeguVorlageKey key = EbeguVorlageKey.getBenutzerHandbuchKeyForRole(userRole);
 		if (key == null) {
 			return null;
